+++ scilab/modules/helptools/src/java/org/scilab/modules/helptools/HTMLDocbookTagConverter.java
@@ -672,8 +672,9 @@ public class HTMLDocbookTagConverter extends DocbookTagConverter implements Temp
         String idAttr    = attrs.get("id");
         String alignAttr = attrs.get("align");   // mixes align and valign imagedata attributes
         String widthAttr = attrs.get("width");
-        String heightAttr= attrs.get("height"); // officially named "depth" as imagedata attribute
+        String heightAttr= attrs.get("height");  // officially named "depth" as imagedata attribute
         String styleAttr = attrs.get("style");
+        String altAttr = attrs.get("alt");       // for example: LaTeX content as text
         boolean addDiv = getGenerationType() != Backend.JAVAHELP || !isLinkedImage();
         final StringBuilder buffer = new StringBuilder(128);
         if (addDiv && alignAttr != null) {
@@ -695,7 +696,11 @@ public class HTMLDocbookTagConverter extends DocbookTagConverter implements Temp
             }
         }
         if (styleAttr != null) {
-            buffer.append("style=\'").append(styleAttr).append("\'");
+            buffer.append("style=\'").append(styleAttr).append("\' ");
+        }
+        if (altAttr != null) {
+            altAttr = altAttr.replaceAll("\'", "&apos;").replaceAll("\"", "&quot;");
+            buffer.append("alt=\'").append(altAttr).append("\'");
         }
         buffer.append("/>\n");
         if (addDiv && alignAttr != null) {